Move some commands to a staff package
