+++ src/main/java/net/cytonic/cytosis/commands/CommandHandler.java
@@ -7,6 +7,7 @@ import net.cytonic.cytosis.commands.defaultMinecraft.GamemodeCommand;
 import net.cytonic.cytosis.commands.defaultMinecraft.TeleportCommand;
 import net.cytonic.cytosis.commands.moderation.*;
 import net.cytonic.cytosis.commands.server.*;
+import net.cytonic.cytosis.commands.staff.*;
 import net.minestom.server.command.CommandManager;
 import net.minestom.server.entity.Player;
 
@@ -69,6 +70,7 @@ public class CommandHandler {
         cm.register(new MsgCommand());
         cm.register(new WarnCommand());
         cm.register(new CooldownCommand(Cytosis.getNetworkCooldownManager()));
+        cm.register(new LoopCommand());
     }
 
     /**
